refactor: move components into individual folders
