+++ src/components/index.js
@@ -1,2 +1,2 @@
-export { default as ZoaButton } from './Button.ce.vue';
-export { default as ZoaTextbox } from './Textbox.ce.vue';
+export { default as ZoaButton } from './inputs/button/Button.ce.vue';
+export { default as ZoaTextbox } from './inputs/textbox/Textbox.ce.vue';
