+++ src/components/inputs/button/button.scss
@@ -1,6 +1,6 @@
-@import 'palette';
-@import 'fonts';
-@import 'vars';
+@import '../../../styles/palette';
+@import '../../../styles/fonts';
+@import '../../../styles/vars';
 
 .button {
   font-family: $header-font;
