+++ src/components/inputs/textbox/textbox.scss
@@ -1,6 +1,6 @@
-@import 'palette';
-@import 'fonts';
-@import 'vars';
+@import '../../../styles/palette';
+@import '../../../styles/fonts';
+@import '../../../styles/vars';
 
 input {
   font-family: $body-font;
